requirements.txt
Django==5.1.3
oracledb==2.4.1
Pillow==10.4.0
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
python-decouple==3.8
pytz==2024.1
openpyxl==3.1.2